Feat/municipality data in API (#572)
* Add municipality climate data and remove old company data file

* Improve type for getErrorSchemas

* Add municipality data to the API

- `GET /municipalities` returns a list of all entries.
- `GET /municipalities/:name` returns a specific entry.

The `municipalityService` is lazy-loading data the first time it is requested, which reduces the startup time and memory usage until the data is actually needed. In addition we also send regular cache headers to the client.

The `climate-data.json` rarely changes and the file is part of this git repo. In the future we also want to move the Python data pipeline here too.

* Update schema to match new English property names